requirements_backup.txt
amqp==5.1.1
aniso8601==9.0.1
asgiref==3.5.2
asttokens==2.1.0
attrs==21.4.0
backcall==0.2.0
billiard==3.6.4.0
boto3==1.23.0
boto3-stubs==1.23.0
botocore==1.26.10
botocore-stubs==1.29.5
celery==5.2.6
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
decorator==5.1.1
Django==3.2.13
django-celery-beat==2.2.1
django-celery-results==2.3.1
django-cors-headers==3.12.0
django-environ==0.8.1
django-extensions==3.1.5
django-filter==21.1
django-graphql-auth==0.3.7
django-graphql-jwt==0.3.4
django-jazzmin==2.6.0
django-storages==1.12.3
django-stubs==1.10.1
django-stubs-ext==0.7.0
django-timezone-field==4.2.3
djangorestframework==3.13.1
djangorestframework-stubs==1.5.0
drf-jwt==1.19.2
executing==1.2.0
factory-boy==3.2.1
Faker==13.11.1
flake8==4.0.1
graphene==3.2.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.4
iniconfig==1.1.1
ipdb==0.13.9
ipython==8.3.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.2.4
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mccabe==0.6.1
mypy==0.942
mypy-extensions==0.4.3
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.0.0
promise==2.3
prompt-toolkit==3.0.32
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.13.0
PyJWT==2.6.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
traitlets==5.5.0
types-awscrt==0.15.2
types-Markdown==3.4.2.1
types-pytz==2022.6.0.1
types-PyYAML==6.0.12.2
types-requests==2.28.11.4
types-urllib3==1.26.25.3
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.1.0
